# openapi.json
{"openapi":"3.1.0","info":{"title":"Brain AG - Backend","version":"0.3.0"},"paths":{"/metrics":{"get":{"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/producers":{"post":{"tags":["producers"],"summary":"Create Producer","operationId":"create_producer_producers_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["producers"],"summary":"List Producers","operationId":"list_producers_producers_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProducerRead"},"title":"Response List Producers Producers Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/producers/{producer_id}":{"get":{"tags":["producers"],"summary":"Get Producer","operationId":"get_producer_producers__producer_id__get","parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["producers"],"summary":"Update Producer","operationId":"update_producer_producers__producer_id__put","parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["producers"],"summary":"Delete Producer","operationId":"delete_producer_producers__producer_id__delete","parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/farms":{"post":{"tags":["farms"],"summary":"Create Farm","operationId":"create_farm_farms_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["farms"],"summary":"List Farms","operationId":"list_farms_farms_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"producer_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Producer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FarmRead"},"title":"Response List Farms Farms Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/farms/{farm_id}":{"get":{"tags":["farms"],"summary":"Get Farm","operationId":"get_farm_farms__farm_id__get","parameters":[{"name":"farm_id","in":"path","required":true,"schema":{"type":"integer","title":"Farm Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["farms"],"summary":"Update Farm","operationId":"update_farm_farms__farm_id__put","parameters":[{"name":"farm_id","in":"path","required":true,"schema":{"type":"integer","title":"Farm Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["farms"],"summary":"Delete Farm","operationId":"delete_farm_farms__farm_id__delete","parameters":[{"name":"farm_id","in":"path","required":true,"schema":{"type":"integer","title":"Farm Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/farm-crops":{"post":{"tags":["farm-crops"],"summary":"Add Farm Crop","operationId":"add_farm_crop_farm_crops_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmCropCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FarmCropRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["farm-crops"],"summary":"List Farm Crops","operationId":"list_farm_crops_farm_crops_get","parameters":[{"name":"farm_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Farm Id"}},{"name":"season","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Season"}},{"name":"crop","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Crop"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FarmCropRead"},"title":"Response List Farm Crops Farm Crops Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/summary":{"get":{"tags":["dashboard"],"summary":"Summary","operationId":"summary_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/pie/state":{"get":{"tags":["dashboard"],"summary":"Pie State","operationId":"pie_state_dashboard_pie_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/pie/crop":{"get":{"tags":["dashboard"],"summary":"Pie Crop","operationId":"pie_crop_dashboard_pie_crop_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/pie/landuse":{"get":{"tags":["dashboard"],"summary":"Pie Landuse","operationId":"pie_landuse_dashboard_pie_landuse_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"FarmCreate":{"properties":{"producer_id":{"type":"integer","title":"Producer Id"},"name":{"type":"string","title":"Name"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"area_total":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Total"},"area_agricultavel":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Agricultavel"},"area_vegetacao":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Vegetacao"}},"type":"object","required":["producer_id","name","area_total","area_agricultavel","area_vegetacao"],"title":"FarmCreate"},"FarmCropCreate":{"properties":{"farm_id":{"type":"integer","title":"Farm Id"},"season":{"type":"string","title":"Season"},"crop":{"type":"string","title":"Crop"}},"type":"object","required":["farm_id","season","crop"],"title":"FarmCropCreate"},"FarmCropRead":{"properties":{"id":{"type":"integer","title":"Id"},"farm_id":{"type":"integer","title":"Farm Id"},"season":{"type":"string","title":"Season"},"crop":{"type":"string","title":"Crop"}},"type":"object","required":["id","farm_id","season","crop"],"title":"FarmCropRead"},"FarmRead":{"properties":{"id":{"type":"integer","title":"Id"},"producer_id":{"type":"integer","title":"Producer Id"},"name":{"type":"string","title":"Name"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"area_total":{"type":"string","title":"Area Total"},"area_agricultavel":{"type":"string","title":"Area Agricultavel"},"area_vegetacao":{"type":"string","title":"Area Vegetacao"}},"type":"object","required":["id","producer_id","name","city","state","area_total","area_agricultavel","area_vegetacao"],"title":"FarmRead"},"FarmUpdate":{"properties":{"name":{"type":"string","title":"Name"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"area_total":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Total"},"area_agricultavel":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Agricultavel"},"area_vegetacao":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Area Vegetacao"}},"type":"object","required":["name","area_total","area_agricultavel","area_vegetacao"],"title":"FarmUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProducerCreate":{"properties":{"cpf_cnpj":{"type":"string","title":"Cpf Cnpj"},"name":{"type":"string","title":"Name"}},"type":"object","required":["cpf_cnpj","name"],"title":"ProducerCreate"},"ProducerRead":{"properties":{"id":{"type":"integer","title":"Id"},"cpf_cnpj":{"type":"string","title":"Cpf Cnpj"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","cpf_cnpj","name"],"title":"ProducerRead"},"ProducerUpdate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"ProducerUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}